backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-session==0.8.0
msal==1.31.1
requests==2.32.3
openai==1.68.0
python-dotenv==1.0.1

# Configuration
PyYAML==6.0.1

# Testing dependencies (optional for production)
pytest==8.1.1
pytest-cov==5.0.0
pytest-mock==3.14.0
Faker==24.4.0
